Move Cache to a separate class
