+++ src/main/java/org/jenkinsci/plugins/rundeck/cache/RundeckJobCache.java
@@ -3,7 +3,7 @@ package org.jenkinsci.plugins.rundeck.cache;
 import org.rundeck.api.RundeckClient;
 import org.rundeck.api.domain.RundeckJob;
 
-public interface IRundeckJobCache {
+public interface RundeckJobCache {
 
     RundeckJob findJobById(final String rundeckJobId, final String rundeckInstanceName, final RundeckClient rundeckInstance);
